third_party/blink/web_tests/external/wpt/xhr/responsetype.any.worker-expected.txt

This is a testharness.js-based test.
Found 7 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Set responseType to "document" when readyState is UNSENT.
  assert_equals: expected "" but got "document"
[FAIL] Set responseType to "document" when readyState is OPENED.
  assert_equals: expected "" but got "document"
[FAIL] Set responseType to "document" when readyState is HEADERS_RECEIVED.
  assert_equals: expected "" but got "document"
[FAIL] Set responseType to "document" when readyState is LOADING.
  Failed to set the 'responseType' property on 'XMLHttpRequest': The response type cannot be set if the object's state is LOADING or DONE.
[FAIL] Set responseType to "document" when readyState is DONE.
  Failed to set the 'responseType' property on 'XMLHttpRequest': The response type cannot be set if the object's state is LOADING or DONE.
[FAIL] Set responseType to "document" when readyState is OPENED and the sync flag is set.
  assert_equals: expected "" but got "document"
[FAIL] Set responseType to "document" when readyState is DONE and the sync flag is set.
  Failed to set the 'responseType' property on 'XMLHttpRequest': The response type cannot be set if the object's state is LOADING or DONE.
Harness: the test ran to completion.
